requirements.txt
pyTelegramBotAPI==3.6.7
PyYAML==5.3.1
PySocks==1.7.1
